Do not shortcut replicator job initialization if replication ID matches
Previously there was an attempt to shortcut some of the job initialization if
replication ID in the job data and the newly computed one matched. However,
that logic was wrong as it skipped over the job data state update.

The effect was that if a job was in a pending state, and it re-initialized, say
when a node restarted, its job data would still indicate it as "pending" until
the next checkpoint. If the job is continuous, and there are no more updates on
the source, the state of the job would stay in "pending" indefinitely.