Fix semantics of total_rows
The `total_rows` field suppose to be a number of documents in the database/view.
See https://docs.couchdb.org/en/stable/api/ddoc/views.html.

When new pagination API was introduced the meaning of `total_rows` field was
changed to number of rows in the query results. The PR reverts this
accidental change.