Retry filter_docs sequentially if the patch exceeds couchjs stack
A document with lots of conflicts can blow up couchjs if the user
calls _changes with a javascript filter and with `style=all_docs` as
this option causes up to fetch all the conflicts.

All leaf revisions of the document are then passed in a single call to
ddoc_prompt, which can fail if there's a lot of them.

In that event, we simply try them sequentially and assemble the
response from each call.

Should be backported to 3.x