Fix flaky test
Waiting for the timeout option to be set means we could still sneak in
and grab the old FDB database handle before fabric2_server updated it in
the application environment.

This new approach just waits until the handle has been updated by
watching the value in the application environment directly.