Minimize conflicts while building views
If a client is continuously updating a database the view engine would
fail to make progress because the last update sequence is continuously
changing and the unbounded changes read would create a conflict to the
maximum versionstamp value used to delineate the end of the changes key
space.

This commits makes two changes to avoid this. First, the current
update_seq is read via snapshot so that it doesn't become a read
conflict if there's an update. This part is fairly straightforward.

The second change reads the changes feed in a snapshot and then manually
adds the read conflict range for the keys that were actually read. This
solves the issue for large view builds while a database is being
updated. However, this does not address the issue when its a single
document being updated repeatedly. Perhaps if we keep track of the last
sequence read and apply that as a limit when updating the view on a
conflict retry?