Remove incorrect MD5 check
If an attachment was stored uncompressed but later is replicated
internally to a node that wants to compress it (based on
content-type), couchdb compares the uncompressed md5 with the
compressed md5 and fails. This breaks eventual consistency between
replicas.

This PR removes the unnecessary MD5 check that is, in these specific
circumstances, always called with mismatched arguments.