Add a .devcontainer configuration for 3.x (#3336)
This PR adds a Dockerfile and associated configuration to enable
developers to quickly provision an environment with all dependencies
installed to work on CouchDB 3.x.

The container configuration also installs the Erlang Language Server
extension. That extension needs a minimal configuration file in the root
of the project in order to find the include files, so I've added that as
well. We could likely iterate and enhance that configuration file
further with linters, dialyzers configurations, etc.

Finally, it allows a developer to set the SpiderMonkey version in an
$SM_VSN environment variable so that we can do a better job of
preserving the simplicity of `./configure; make` inside the container.