Ignore unchecked JWT claims
Previously, if a JWT claim was present, it was validated regardless of
whether it was required.

However, according to the spec [1]:

"all claims that are not understood by implementations MUST be ignored"

which we interpret to mean that we should not attempt to validate
claims we don't require.

With this change, only claims listed in required checks are validated.

[1] https://tools.ietf.org/html/rfc7519#section-4