Consistent view emits using indexer's GRVs and committed versionstamps
View indexer saves both the GRV it used during the view update, and
the committed versionstamp in the couch_jobs job data section. Then,
the view reader uses those versionstamps to emit a consistent snapshot
of the view.

  * The committed versionstamp ensures that the view results can be
    emitted even if the view gets updated between the time the view
    finished and the reader gets notified.

  * The indexer GRV ensures that the view will emit the same doc
    revisions in case when include_docs=true option is used as what it
    read during the time it indexed the data.

The view reader uses those two versionstamps only if it initiates the
view build itself and then waits for it to build, to ensure that it
doesn't operate on stale GRVs.

Because the committed version is only available after the main
transaction commits, during view indexing finalization there is now a
separate transaction which runs at the end which reads the committed
version then marks the view as `finished`.

Since included docs have to be read at the indexer's GRV version, and,
that version is different than the committed version, those documents
are loaded in from a separate process.

There are a few complications introduced by this commit:

   * The versionstamps, especially the indexer GRV ones, may become
     stale (older than 5 seconds) quickly and start throwing 1007
     (transaction_too_old) errors. This could be mitigated by forcing
     the indexer to commit after a shorter interval (1-2 seconds).

   * There is some fragility introduced in respect to how included
     docs are loaded in a separate process. When that crashes or does
     timeout it maybe throw a new type of un-expected error that we
     don't catch properly.