Improve search for FDB cluster files
This patch introduces a handful of related improvments to the code that
chooses the FoundationDB cluster file. We now check for a cluster file
in up to 4 locations, in priority order:

1. The value of `[erlfdb] cluster_file` in the server config
2. The value of the FDB_CLUSTER_FILE environment variable
3. /usr/local/etc/foundationdb/fdb.cluster (MacOS package default)
4. /etc/foundationdb/fdb.cluster (Linux packages default)

The second and fourth locations are new. We also check that the file
has appropriate access permissions, namely that is RW accessible by the
user running CouchDB. Finally, we add some more detailed logging to
help users debug any problems in this department on startup.

If a location is specified using either one of the first two "custom"
approaches and there is a problem with that location, CouchDB will
abort rather than try any further option. This avoids a situation where
a user intended to override the default cluster file but CouchDB starts
up with the default one because e.g. the user failed to actually
install the new file with the correct permissions.