Clean up config files
Remove default.ini and local.ini sections related to applications and
functionality which is no longer there.

One odd change is the commented out`partitioned` flag. The reason is because
the "flags" logic is kept, and seems to be functional, however partitioned flag
is not functional. `couch_flags` loads flags using `list_to_existing_atom/1`
and since `partitioned` is not an existing atom it continuously emits crash
logs from the `_up` point related to that.