Remove most of the functionality from couch_server
Remove all the code related to opening and caching of databases.

However, couch_server did a few other things such as:

 * Parse and maintain the CouchDB version

 * Return the server "uuid" value

 * The gen_server was monitoring config updates and hashing admin passwords
   when they were updated

It was a 50/50 decision to move that functionality out to other modules
completely or keep it where it is. Since it wasn't just a single thing, and the
overall PR was getting rather large, opted to pair the exisiting code to the
minimum, and then later we can do another round of cleanup and find a better
place for those features.