Add structured logging reports via new Erlang 21 logger (#3526)
* Drop support for Erlang < 21

The new logger only works in Erlang 21+, so we can start simplifying
the codebase by removing the macro that provides support for retrieving
stack traces the old way.

* Produce structured logging reports

We use the new logger macros to generate log events alongside the
existing couch_log system. This commit does not introduce any handlers
for the new events; that will come later.