Handle the case when md5 field is undefined
Previously the default value for md5 field was `<<>>`. This value changed to
undefined when we switch to using maps instead of erlang records.
The change break the `couch_att:to_json/4` funciton because `base64:encode/1`
cannot handle atom `undefined`.

The issue happens for inline attachments only this is why it wasn't discovered
earlier.