Use the last commit result even when there are intermediate retries
Previously, if a transaction got a `commit_unknown_result`, during the next
successful attempt that result would be returned to the user. However, if the
next attempt was another retryable error, then the committed result was ignored
and the whole transaction would be applied again. This could result in document
update transactions conflicting with themselves as described in issue
https://github.com/apache/couchdb/issues/3560

To prevent that from happening we remember that there was an
`commit_unknown_result` error during the course of any previous retries and try
to return that result.