Discard a payload on a delete attachment request
While including a payload within a DELETE request is not forbidden by RFC7231
its presence on a delete attachment request leaves a mochiweb acceptor
in a semi-opened state since mochiweb's using lazy load for the request bodies.
This makes a next immediate request to the same acceptor to hung
until previous request's receive timeout.

This PR adds a step to explicitly "drain" and discard an entity body on a
delete attachment request to prevent that.